Refactor airspace out of core libraries
