+++ src/core/mod.rs
@@ -1,4 +1,3 @@
-pub mod airspace;
 pub mod central_disk_logger;
 pub mod ingestor;
 pub mod parser;
