+++ src/gui/app.rs
@@ -1,7 +1,7 @@
 use eframe::{egui, epaint};
 use walkers;
 
-use crate::core::airspace::AirspaceViewer;
+use crate::airspace::AirspaceViewer;
 use crate::core::parser::Aircraft;
 use crate::gui::constants::AIRCRAFT_REFERENCE_SHAPE;
