+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod airspace;
 pub mod cli;
 pub mod core;
 pub mod ext;
